pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "twotrim"
version = "0.1.0"
description = "Universal token compression fabric for LLM applications"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "TwoTrim Contributors"},
]
keywords = ["llm", "compression", "tokens", "proxy", "middleware", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "tiktoken>=0.7.0",
    "tokenizers>=0.19.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "prometheus-client>=0.20.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "structlog>=24.1.0",
    "openai>=1.30.0",
    "aiosqlite>=0.20.0",
    "xxhash>=3.4.0",
    "regex>=2024.4.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.2.0",
]
redis = [
    "redis[hiredis]>=5.0.0",
]
vllm = [
    "vllm>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
evaluate = [
    "datasets>=2.14.0,<2.20.0",
    "evaluate>=0.4.0",
    "rouge_score>=0.1.2",
]

[project.scripts]
twotrim = "twotrim.cli.main:cli"

[project.urls]
Homepage = "https://github.com/twotrim/twotrim"
Repository = "https://github.com/twotrim/twotrim"
Documentation = "https://github.com/twotrim/twotrim/tree/main/docs"
Issues = "https://github.com/twotrim/twotrim/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/twotrim"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
